nuX: Removing printer calls, changing to printf.
Removed Printer:: use, changed to printf statements which are GPU compatible.
Additionally, moved the error print statements to after the loop to avoid printing during the loop execution.
Print information about the point is not functional yet as the point info
is only available during the loop execution, not after it.